requirements.txt
streamlit>=1.54,<2
python-docx>=1.2,<2
pandas>=2.3,<3
supabase>=2.15,<3
